% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alpha0}
\alias{alpha0}
\title{Initialize Alpha}
\usage{
alpha0(Zs, s)
}
\arguments{
\item{Zs}{Surrogate design matrix}

\item{s}{Surrogate outcome}
}
\description{
Initialize \eqn{\alpha} for multivariate outcome model.
}
