% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BNRalpha}
\alias{BNRalpha}
\title{Bivariate Normal, Update for Alpha}
\usage{
BNRalpha(t, Zt, Zs, A, b, a0, L)
}
\arguments{
\item{t}{Target outcome}

\item{Zt}{Target design}

\item{Zs}{Surrogate design}

\item{A}{Surrogate design inner product}

\item{b}{Current beta}

\item{a0}{Initial alpha}

\item{L}{Current precision}
}
\description{
Parameter update for bivariate outcome model.
}
