% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BNRbeta}
\alias{BNRbeta}
\title{Bivariate Normal, Update for Beta}
\usage{
BNRbeta(s, Zt, B, Zs, b0, a, L)
}
\arguments{
\item{s}{Surrogate outcome}

\item{Zt}{Target design}

\item{B}{Target design inner product.}

\item{Zs}{Surrogate design}

\item{b0}{Initial beta}

\item{a}{Current alpha}

\item{L}{Current precision.}
}
\description{
Parameter update \eqn{\beta} in the bivariate outcome model.
}
