% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{coef.mnr}
\alias{coef.mnr}
\title{Extract Coefficients from Multivariate Regression Model}
\usage{
\method{coef}{mnr}(object, ..., type = "Beta")
}
\arguments{
\item{object}{A \code{mnr} object.}

\item{...}{Unused.}

\item{type}{Either "Beta" or "Alpha".}
}
\description{
Returns the estimated regression coefficients from a fitted \code{mnr} model. By
default, returns \eqn{\beta}, the regression coefficients for the target
outcome.
}
