% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{coef.mvr}
\alias{coef.mvr}
\title{Extract Coefficients from Multivariate Regression Model}
\usage{
\method{coef}{mvr}(object, ..., type = "Beta")
}
\arguments{
\item{object}{A \code{mvr} object.}

\item{...}{Unused.}

\item{type}{Either Alpha or Beta.}
}
\description{
Extract Coefficients from Multivariate Regression Model
}
