% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiTest.R
\name{constrXi}
\alias{constrXi}
\title{Construct Overall Surrogate Design}
\usage{
constrXi(Ls, parallel)
}
\arguments{
\item{Ls}{List of \eqn{k-1} model matrices}

\item{parallel}{If true, assumes a parallel coefficient design, where each 
surrogate outcome has the same regression parameters.}
}
\description{
Assembles the overall surrogate design from a list of design matrices for the
individual surrogate outcomes.
}
