% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{D.bnr}
\alias{D.bnr}
\title{Simulated Data with Bivariate Normal Outcomes}
\format{A list of vectors and matrices. 
\describe{ 
   \item{y.t}{Numeric vector containing the target outcome.}
   \item{y.s}{Numeric vector containing the surrogate outcome.}
   \item{D.t}{Design matrix for the target outcome.}
   \item{D.s}{Design matrix for the surrogate outcome.}
   \item{Beta}{Regression coefficients used to generate the target outcome.}
   \item{Alpha}{Regression coefficients used to generate the surrogate outcome.}
}}
\usage{
D.bnr
}
\description{
Example data for 1000 subjects. The outcomes are bivariate normal with unit
variances and correlation \eqn{\rho=0.5}. Each outcome depends on a design matrix containing three independent 
N(0,1) covariates. \code{Beta} contains regression coefficients for the target outcome, and \code{Alpha} contains
regression coefficients for the secondary ("surrogate") outcome.
}
\keyword{datasets}
