% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{D.mnr}
\alias{D.mnr}
\title{Simulated Data with Triivariate Normal Outcomes}
\format{A list of vectors and matrices. 
\describe{ 
  \item{y.t}{Numeric vector containing the target outcome.} 
  \item{y.s}{Numeric matrix containing the surrogate outcomes.} 
  \item{Z.t}{Model matrix for the target outcome.} 
  \item{L.s}{List of model matrices for the surrogate outcomes.} 
  \item{Beta}{Regression coefficients used to generate the target outcome.} 
  \item{Alpha}{Regression coefficients used to generate the surrogate 
  outcome.} }}
\usage{
D.mnr
}
\description{
Example data for 1000 subjects. The outcomes are trivariate normal with unit 
variances and an exchangeable correlation structure, \eqn{\rho=0.5}. The 
target and first surrogate outcomes each depend on three independent N(0,1)
covariates, while the second surrogate outcome depends on four such
covariates. \code{Beta} contains regression coefficients for the target
outcome, and \code{Alpha} contains regression coefficients for the secondary
("surrogate") outcomes.
}
\keyword{datasets}
