% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMatrixOps.R
\name{effInfo}
\alias{effInfo}
\title{Efficient Information}
\usage{
effInfo(L, M)
}
\arguments{
\item{L}{Logical vector indicating the components of interest}

\item{M}{Fitted model}
}
\description{
Calculate efficient information for a subset of components if \eqn{\beta}.
}
