% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiFit.R
\name{fit.mnr}
\alias{fit.mnr}
\title{Fit Multivariate Outcome Model}
\usage{
fit.mnr(yt, Ys, Zt, Zs, maxit = 10, eps = 1e-06, report = T)
}
\arguments{
\item{yt}{Target outcome vector.}

\item{Ys}{Surrogate outcome matrix.}

\item{Zt}{Target model matrix.}

\item{Zs}{Surrogate model matrix.}

\item{maxit}{Maximum number of parameter updates.}

\item{eps}{Minimum acceptable improvement in log likelihood.}

\item{report}{Report fitting progress? Default is FALSE.}
}
\value{
An object of class \code{mnr} containing the fitted regression 
  parameters, the estimated target-surrogate covariance matrix, the 
  regression parameter information matrix, and the estimated residuals.
}
\description{
Fits the multivariate outcome regression model.
}
\details{
The target and surrogate model matrices are expected in numeric format. 
Expand factors and interactions in advance. If an intercept is required, 
include a vector of constants. The inputs should contain no missing values.
}
