% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fitNorm}
\alias{fitNorm}
\title{Univariate OLS model.}
\usage{
fitNorm(y, Z)
}
\arguments{
\item{y}{Outcome.}

\item{Z}{Model matrix.}
}
\value{
A list containing the following:
\item{Beta}{Regression coefficient.}
\item{Tau}{Outcome variance.}
\item{Ibb}{Information matrix for beta.}
\item{Resid}{Outcome residuals.}
}
\description{
Fits the standard OLS model.
}
