% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{G}
\alias{G}
\title{Simulated Genotypes}
\format{A numeric matrix with 1000 rows and 1000 columns.}
\usage{
G
}
\description{
Simulated genotypes for 1000 subjects at 1000 loci. Neither the target nor
the surrogate phenotypes depend on these genotypes. The function
\code{\link{rScore.bnr}} provides a way to rapidly test the target phenotype
for association with each locus.
}
\keyword{datasets}
