% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{infoMNR}
\alias{infoMNR}
\title{Information}
\usage{
infoMNR(n, A, B, L)
}
\arguments{
\item{n}{Observations}

\item{A}{Left hand matrix}

\item{B}{Right hand matrix}

\item{L}{Precision}
}
\description{
Calculate \eqn{A'(I_{n\times n}\otimes L)B}, the general form of information matrices
for the multivariate outcome model.
}
