% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\docType{methods}
\name{show,mnr-method}
\alias{show,mnr-method}
\title{Show for Multivariate Regression Model}
\usage{
\S4method{show}{mnr}(object)
}
\arguments{
\item{object}{A \code{mnr} object.}
}
\description{
Show for Multivariate Regression Model
}
