% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MNRalpha}
\alias{MNRalpha}
\title{Multivariate Normal, Update for Alpha}
\usage{
MNRalpha(t, s, Zt, Zs, b, L)
}
\arguments{
\item{t}{Target outcome}

\item{s}{Surrogate outcome}

\item{Zt}{Target design}

\item{Zs}{Surrogate design}

\item{b}{Current beta}

\item{L}{Current precision}

\item{a0}{Initial alpha}
}
\description{
Parameter update for multivariate outcome model.
}
