% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MNRbeta}
\alias{MNRbeta}
\title{Multivariate Normal, Update for Beta}
\usage{
MNRbeta(s, Zt, B, Zs, b0, a, L)
}
\arguments{
\item{s}{Surrogate outcome}

\item{Zt}{Target design}

\item{B}{Target design inner product.}

\item{Zs}{Surrogate design}

\item{b0}{Initial beta}

\item{a}{Current alpha}

\item{L}{Current precision.}
}
\description{
Parameter update for \eqn{\beta} in the multivariate outcome model.
}
