% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\docType{class}
\name{mvr-class}
\alias{mvr-class}
\title{Multivariate Regression Model}
\description{
Multivariate Regression Model
}
\section{Slots}{

\describe{
\item{\code{Coefficients}}{Regression coefficients.}

\item{\code{Covariance}}{Outcome covariance matrix.}

\item{\code{Information}}{Information.}

\item{\code{Residuals}}{Phenotypic residuals.}
}}

