% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{print.mnr}
\alias{print.mnr}
\title{Print for Multivariate Regression Model}
\usage{
\method{print}{mnr}(x, ...)
}
\arguments{
\item{x}{A \code{mnr} object.}

\item{...}{Unused.}
}
\description{
Print method for objects of class \code{mnr}.
}
