% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{print.mvr}
\alias{print.mvr}
\title{Print for Multivariate Regression Model}
\usage{
\method{print}{mvr}(x, ...)
}
\arguments{
\item{x}{A \code{mvr} object.}

\item{...}{Unused.}
}
\description{
Print for Multivariate Regression Model
}
