% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{residuals.mvr}
\alias{residuals.mvr}
\title{Extract Residuals from Multivariate Regression Model}
\usage{
\method{residuals}{mvr}(object, ..., type = "Target")
}
\arguments{
\item{object}{A \code{mvr} object.}

\item{...}{Unused.}

\item{type}{Either Surrogate or Target.}
}
\description{
Extract Residuals from Multivariate Regression Model
}
