% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{rMNR}
\alias{rMNR}
\title{Simulate Multivariate Normal Data}
\usage{
rMNR(X, b, S)
}
\arguments{
\item{X}{List of design matrices, one for each outcome.}

\item{b}{List of regression coefficients, one for each outcome.}

\item{S}{Outcome covariance structure.}
}
\value{
Numeric \eqn{n\times k} matrix, where \eqn{n} is the number of rows in each
design matrix, and \eqn{k} is the number of rows in the covariance structure.
}
\description{
Function to simulate outcomes from a multivariate normal regression model.
}
\examples{
\dontrun{
set.seed(100);
# Observations
n = 1e3;
## Design matrices
X1 = cbind(1,matrix(rnorm(2*n),nrow=n));
colnames(X1) = c("int",paste0("x0",seq(1:2)));
X2 = cbind(1,matrix(rnorm(3*n),nrow=n));
colnames(X2) = c("int",paste0("x1",seq(1:3)));
X3 = cbind(1,matrix(rnorm(4*n),nrow=n));
colnames(X3) = c("int",paste0("x2",seq(1:4)));
X = list(X1,X2,X3);
# Target Parameter
b1 = c(-1,0.1,-0.1);
b2 = c(1,-0.1,0.1,0);
b3 = c(0,0.1,-0.1,0.1,-0.1);
b = list(b1,b2,b3);
# Exchangeable covariance structure
S = array(0.5,dim=c(3,3)) + 0.5*diag(3);
# Generate data
Y = rMNR(X=X,b=b,S=S);
}
}
