% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bivariate.R
\name{rScore.bnr}
\alias{rScore.bnr}
\title{Bivariate Regression Score Test}
\usage{
rScore.bnr(y.t, y.s, X1, X2, Z.s, maxit = 10, eps = 1e-06)
}
\arguments{
\item{y.t}{Target outcome.}

\item{y.s}{Surrogate outcome.}

\item{X1}{Numeric matrix of covariates for the target outcome whose
regression coefficients are constrained under the null.}

\item{X2}{Numeric matrix of covariates for the target outcome whose
regression coefficients are unconstrained under the null.}

\item{Z.s}{Numeric model matrix for the surrogate outcome.}

\item{maxit}{Maximum number of parameter updates.}

\item{eps}{Minimum acceptable improvement in log likelihood.}
}
\description{
Individually tests \eqn{H_{0}:\beta_{j}=0} for each column of \code{X1}, 
adjusting for \code{X2}.
}
