% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rScore.R
\name{rScore.mnr}
\alias{rScore.mnr}
\title{Multivariate Regression Repeated Score Test}
\usage{
rScore.mnr(Y, j = 1, G, X, maxit = 100, eps = 1e-08, report = F,
  parallel = F)
}
\arguments{
\item{Y}{Outcome matrix.}

\item{j}{Column number of the outcome of interest. By default, \code{j=1}.}

\item{G}{Numeric matrix of covariates for the target outcome whose regression
coefficients are zero under the null.}

\item{X}{List of model matrices, one for each outcome.}

\item{maxit}{Maximum number of parameter updates.}

\item{eps}{Minimum acceptable improvement in log likelihood.}

\item{report}{Report model fitting progress? Default is FALSE.}

\item{parallel}{Run association testing in parallel? Must register parallel
backend first.}
}
\value{
A numeric vector of p-values, one for each column of \code{G}.
}
\description{
Individually tests each column of \code{G} for association with the target 
outcome. \code{G} may contain missing elements, although the remaining model
matrices may not. \code{rScore.mnr} accelerates association testing by
recycling the same null model for each hypothesis test.
}
\examples{
\dontrun{
# See rMNR or vignette for data generation
# Genotype matrix
G = replicate(2000,rbinom(n=1000,size=2,prob=0.25));
storage.mode(G) = "numeric";
# Introduce missingness
G[sort(sample(length(G),size=0.01*length(G)))] = NA;
# Repeated Score test
R = rScore.mnr(Y=Y,G=G,X=X,report=T);
# Estimated size
mean(R<=0.05);
}
}
