% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bivariate.R
\name{Score.bnr}
\alias{Score.bnr}
\title{Bivariate Regression Score Test}
\usage{
Score.bnr(y.t, y.s, Z.t, Z.s, L, b0, maxit = 10, eps = 1e-06)
}
\arguments{
\item{y.t}{Target outcome.}

\item{y.s}{Surrogate outcome.}

\item{Z.t}{Numeric model matrix for the target outcome.}

\item{Z.s}{Numeric model matrix for the surrogate outcome.}

\item{L}{Logical vector, with as many entires as columns in the target 
design, indicating which columns have fixed coefficients under the null.}

\item{b0}{Value of the regression coefficient for the selected columns under 
the null. Defaults to zero.}

\item{maxit}{Maximum number of parameter updates.}

\item{eps}{Minimum acceptable improvement in log likelihood.}
}
\description{
Score test of \eqn{H_{0}:\beta_{1}=\beta_{10}} in the bivariate outcome
model.
}
\examples{
y.t = D.bnr$y.t;
y.s = D.bnr$y.s;
Z.t = D.bnr$Z.t;
Z.s = D.bnr$Z.s;
# Test for an overall effect
# Note: Null model should contain the intercept
Score.bnr(y.t,y.s,Z.t,Z.s,L=c(FALSE,TRUE,TRUE,TRUE));
# Test for the effect of \\eqn{\\beta_{3}}
Score.bnr(y.t,y.s,Z.t,Z.s,L=c(FALSE,FALSE,FALSE,TRUE));
}
