% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Score.R
\name{Score.mnr}
\alias{Score.mnr}
\title{Multivariate Regression Score Test}
\usage{
Score.mnr(Y, j = 1, X, L, b10 = NULL, maxit = 100, eps = 1e-08,
  report = F)
}
\arguments{
\item{Y}{Outcome matrix.}

\item{j}{Column number of the outcome of interest. By default, \code{j=1}.}

\item{X}{List of model matrices, one for each outcome.}

\item{L}{Logical vector, with as many entires as columns in the target
design matrix, indicating which columns design are fixed under the null.}

\item{b10}{Value of the regression coefficient for the selected columns under
the null. Defaults to zero.}

\item{maxit}{Maximum number of parameter updates.}

\item{eps}{Minimum acceptable improvement in log likelihood.}

\item{report}{Report model fitting progress? Default is FALSE.}
}
\value{
A numeric vector containing the score statistic, degrees of freedom,
  and p-value.
}
\description{
For an outcome of interest among \code{Y}, tests the hypothesis that a subset
of the regression coefficients for that outcome are fixed at the reference 
value \code{b10}. In particular, suppose \eqn{\beta} denotes the regression 
coefficient for the target outcome. Partition 
\eqn{\beta=(\beta_{1},\beta_{2})}. \code{Score.bnem} performs a score test of
\eqn{H_{0}:\beta_{1}=\beta_{10}}.
}
\examples{
\dontrun{
# See rMNR or vignette for data generation
# Test b13 = 0, which is FALSE
Score.mnr(Y=Y,j=1,X=X,L=c(F,F,T));
# Test b24 = 0, which is TRUE
Score.mnr(Y=Y,j=2,X=X,L=c(F,F,F,T));
# Test b32 = ... = b35 = 0, which is FALSE
Score.mnr(Y=Y,j=3,X=X,L=c(F,T,T,T,T));
# Test b32 = b34 = 0.1, which is TRUE
Score.mnr(Y=Y,j=3,X=X,b10=c(0.1,0.1),L=c(F,T,F,T,F));
}
}
