% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sKroneckerP}
\alias{sKroneckerP}
\title{Sparse Kronecker Product}
\usage{
sKroneckerP(n, A)
}
\arguments{
\item{n}{Dimension of identity matrix}

\item{A}{Numeric matrix}
}
\description{
Forms the product \eqn{I_{n\times n}\otimes A}.
}
