% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updateBVR}
\alias{updateBVR}
\title{Bivariate parameter update}
\usage{
updateBVR(yt, ys, Zt, Zs, At, As, b0, a0, s0)
}
\arguments{
\item{yt}{Target outcome}

\item{ys}{Surrogate outcome}

\item{Zt}{Target design}

\item{Zs}{Surrogate design}

\item{At}{Target inc. proj.}

\item{As}{Surrogate inc. proj.}

\item{b0}{Current beta}

\item{a0}{Current alpha}

\item{s0}{Current sigma}
}
\description{
Bivariate parameter update
}
