% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updateMNR}
\alias{updateMNR}
\title{Multivariate Parameter Update}
\usage{
updateMNR(t, s, X, Xi, B, b0, a0, s0)
}
\arguments{
\item{t}{Target outcome}

\item{s}{Surrogate outcome vector}

\item{X}{Target design}

\item{Xi}{Surrogate design}

\item{B}{Target inc. proj.}

\item{b0}{Current beta}

\item{a0}{Current alpha}

\item{s0}{Current sigma}
}
\description{
Parameter update for multivariate outcome model.
}
