% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{vcov.mnr}
\alias{vcov.mnr}
\title{Extract Covariance Matrix from Multivariate Regression Model}
\usage{
\method{vcov}{mnr}(object, ..., type = "Information", inv = F)
}
\arguments{
\item{object}{A \code{mnr} object.}

\item{...}{Unused.}

\item{type}{Either "Information" or "Outcome". Default is "Information".}

\item{inv}{Invert information matrix? Default is FALSE.}
}
\value{
A numeric matrix.
}
\description{
Returns estimated covariance matrices from a fitted \code{mnr} model. Specify
"Regression" for the information matrix of the regression parameters. Specify
"Outcome" for the outcome covariance matrix.
}
