% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{vcov.mvr}
\alias{vcov.mvr}
\title{Extract Covariance Matrix from Multivariate Regression Model}
\usage{
\method{vcov}{mvr}(object, ..., type = "Regression", inv = T)
}
\arguments{
\item{object}{A \code{mvr} object.}

\item{...}{Unused.}

\item{type}{Either Outcome or Regression. Default is Regression.}

\item{inv}{Invert information matrix? Default is TRUE.}
}
\description{
Returns the either the estimated covariance matrix of the outcome,
or the information matrix of the regression coefficients.
}
