% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tests.R
\name{Wald.bnr}
\alias{Wald.bnr}
\title{Bivariate Regression Wald Test}
\usage{
Wald.bnr(y.t, y.s, D.t, D.s, L, b0, maxit = 10, eps = 1e-06)
}
\arguments{
\item{y.t}{Target outcome.}

\item{y.s}{Surrogate outcome.}

\item{D.t}{Target design matrix.}

\item{D.s}{Surrogate design matrix.}

\item{L}{Logical vector, with as many entires as columns in the target
design, indicating which columns of the target design are fixed under the
null.}

\item{b0}{Value of the regression coefficient for the selected columns under
the null. Defaults to zero.}

\item{maxit}{Maximum number of parameter updates.}

\item{eps}{Minimum acceptable improvement in log likelihood.}
}
\description{
Wald test of \eqn{H_{0}:\beta_{1}=\beta_{10}} in the bivariate outcome model.
}
\examples{
y.t = D.bnr$y.t;
y.s = D.bnr$y.s;
D.t = D.bnr$D.t;
D.s = D.bnr$D.s;
# Test for an overall effect
Wald.bnr(y.t,y.s,D.t,D.s,L=c(TRUE,TRUE,TRUE));
# Test for the effect of \\eqn{\\beta_{3}}
Wald.bnr(y.t,y.s,D.t,D.s,L=c(FALSE,FALSE,TRUE));
}
