% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{coef.mnr}
\alias{coef.mnr}
\title{Extract Coefficients from Multivariate Regression Model}
\usage{
\method{coef}{mnr}(object, ..., type = "Beta")
}
\arguments{
\item{object}{A \code{mnr} object.}

\item{...}{Unused.}

\item{type}{Either Alpha or Beta.}
}
\description{
Extract Coefficients from Multivariate Regression Model
}
