% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{coef.mnr}
\alias{coef.mnr}
\title{Extract Coefficients from Multivariate Regression Model}
\usage{
\method{coef}{mnr}(object, ..., outcome = NULL)
}
\arguments{
\item{object}{A \code{mnr} object.}

\item{...}{Unused.}

\item{outcome}{Select an outcome from among}
}
\value{
A data.frame containing estimated coefficients
}
\description{
Returns the estimated regression coefficients from a fitted \code{mnr} model.
}
