% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multivariate.R
\name{constrXi}
\alias{constrXi}
\title{Construct Overall Surrogate Design}
\usage{
constrXi(L.s, parallel)
}
\arguments{
\item{L.s}{List of \eqn{k-1} model matrices}

\item{parallel}{If true, assumes a parallel coefficient design, where each 
secondary outcome has the same regression parameters.}
}
\description{
Assembles the overall surrogate design from the design matrices for the 
individual surrogate outcomes.
}
