% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tests.R
\name{constrXi}
\alias{constrXi}
\title{Construct Overall Surrogate Design}
\usage{
constrXi(D.s, parallel)
}
\arguments{
\item{D.s}{List of \eqn{k-1} design matrices}

\item{parallel}{If true, assumes a parallel coefficient design, where each 
secondary outcome has the same regression parameters.}
}
\description{
Construct Overall Surrogate Design
}
