% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiFit.R
\name{fit.bnr}
\alias{fit.bnr}
\title{Fit Bivariate Outcome Model}
\usage{
fit.bnr(yt, ys, Zt, Zs, maxit = 10, eps = 1e-06, report = T)
}
\arguments{
\item{yt}{Target outcome vector.}

\item{ys}{Surrogate outcome vector.}

\item{Zt}{Target model matrix.}

\item{Zs}{Surrogate model matrix.}

\item{maxit}{Maximum number of parameter updates.}

\item{eps}{Minimum acceptable improvement in log likelihood.}

\item{report}{Report number of iterations?}
}
\description{
Fit the bivariate outcome regression model.
}
\details{
The target and surrogate model matrices are expected in numeric format.
Expand factors and interactions in advance. If an intercept is required,
include a vector of constants.
}
