% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method.R
\name{fit.bnr}
\alias{fit.bnr}
\title{Fit Bivariate Outcome Model}
\usage{
fit.bnr(y.t, y.s, Z.t, Z.s, maxit = 10, eps = 1e-06, report = T)
}
\arguments{
\item{y.t}{Target outcome vector.}

\item{y.s}{Surrogate outcome vector.}

\item{Z.t}{Target model matrix.}

\item{Z.s}{Surrogate model matrix.}

\item{maxit}{Maximum number of parameter updates.}

\item{eps}{Minimum acceptable improvement in log likelihood.}

\item{report}{Report number of iterations?}
}
\description{
Fit the bivariate outcome regression model.
}
