% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit.R
\name{fit.mnr}
\alias{fit.mnr}
\title{Fit Multivariate Outcome Model}
\usage{
fit.mnr(Y, X, sig = 0.05, maxit = 10, eps = 1e-06, report = T)
}
\arguments{
\item{Y}{Outcome matrix.}

\item{X}{List of model matrices, one for each outcome.}

\item{sig}{Significance level, for confidence intervals.}

\item{maxit}{Maximum number of parameter updates.}

\item{eps}{Minimum acceptable improvement in log likelihood.}

\item{report}{Report fitting progress? Default is FALSE.}
}
\value{
An object of class \code{mnr} containing the estimated regression 
  parameters, covariance matrix, the information matrix for regression
  parameters, and the residuals.
}
\description{
Fits a regression model in which a multivariate normal random vector is 
observed for each subject. Regression models are specified using a list of
numeric matrices, one for each column of \code{Y}.
}
\examples{
\dontrun{
# See rMNR or vignette for data generation
M = fit.mnr(Y=Y,X=X,eps=1e-8);
# Coefficients
coef(M);
# Outcome covariance
vcov(M,type="Outcome");
# Information matrix
vcov(M,type="Information");
# Residuals
resid(M);
}
}
