% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\docType{class}
\name{mnr-class}
\alias{mnr-class}
\title{Multivariate Regression Model}
\description{
Defines the object class returned by \code{\link{fit.bnr}} and \code{\link{fit.mnr}}.
}
\section{Slots}{

\describe{
\item{\code{Coefficients}}{Regression coefficients.}

\item{\code{Covariance}}{Outcome covariance matrix.}

\item{\code{Information}}{Information.}

\item{\code{Residuals}}{Phenotypic residuals.}
}}

