% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{residuals.mnr}
\alias{residuals.mnr}
\title{Extract Residuals from Multivariate Regression Model}
\usage{
\method{residuals}{mnr}(object, ...)
}
\arguments{
\item{object}{A \code{mnr} object.}

\item{...}{Unused.}
}
\value{
A numeric matrix of residuals.
}
\description{
Returns the estimated residuals from a fitted \code{mnr} model.
}
