% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{residuals.mnr}
\alias{residuals.mnr}
\title{Extract Residuals from Multivariate Regression Model}
\usage{
\method{residuals}{mnr}(object, ..., type = "Target")
}
\arguments{
\item{object}{A \code{mnr} object.}

\item{...}{Unused.}

\item{type}{Either Surrogate or Target.}
}
\description{
Extract Residuals from Multivariate Regression Model
}
