% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiTest.R
\name{rScore.bnr}
\alias{rScore.bnr}
\title{Repeated Bivariate Regression Score Test}
\usage{
rScore.bnr(yt, ys, Zt1, Zt2, Zs, maxit = 100, eps = 1e-06, report = F)
}
\arguments{
\item{yt}{Target outcome vector.}

\item{ys}{Surrogate outcome vector.}

\item{Zt1}{Numeric matrix of covariates for the target outcome whose 
regression coefficients are zero under the null.}

\item{Zt2}{Numeric matrix of covariates for the target outcome whose 
regression coefficients are unconstrained, i.e. estimated, under the null.}

\item{Zs}{Numeric model matrix for the surrogate outcome.}

\item{maxit}{Maximum number of parameter updates.}

\item{eps}{Minimum acceptable improvement in log likelihood.}

\item{report}{Report model fitting progress? Default is FALSE.}
}
\value{
A numeric vector of p-values, one for each column of \code{Zt1}. The
  p-values are estimated based on the chi-square distribution with one degree
  of freedom.
}
\description{
Individually tests \eqn{H_{0}:\beta=0} for each column of \code{Zt1}, 
adjusting for \code{Zt2}. Missing values are permitted in \code{Zt1}, 
though not in the other model matrices.
}
\examples{
yt = D.bnr$yt;
ys = D.bnr$ys;
Zt2 = D.bnr$Zt;
Zs = D.bnr$Zs;
Zt1 = G;
# Test each column of G for association with the target outcome
R = rScore.bnr(yt,ys,Zt1,Zt2,Zs);
}
