% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiTest.R
\name{Score.bnr}
\alias{Score.bnr}
\title{Bivariate Regression Score Test}
\usage{
Score.bnr(yt, ys, Zt, Zs, L, b10, maxit = 10, eps = 1e-06, report = F)
}
\arguments{
\item{yt}{Target outcome.}

\item{ys}{Surrogate outcome.}

\item{Zt}{Numeric model matrix for the target outcome.}

\item{Zs}{Numeric model matrix for the surrogate outcome.}

\item{L}{Logical vector, with as many entires as columns in the target 
design, indicating which columns have fixed coefficients under the null.}

\item{b10}{Value of the regression coefficient for the selected columns under 
the null. Defaults to zero.}

\item{maxit}{Maximum number of parameter updates.}

\item{eps}{Minimum acceptable improvement in log likelihood.}

\item{report}{Report model fitting progress? Default is FALSE.}
}
\value{
A numeric vector containing the score statistic, the degrees of
  freedom, and a p-value estimated using the chi-square distribution.
}
\description{
Score test of \eqn{H_{0}:\beta_{1}=\beta_{10}} in the bivariate outcome 
model. The score test is specified using a logical vector \code{L}, with as
many entries as columns in the target model matrix \code{Zt}. The values of
\code{L} set to \code{T} are fixed at \eqn{\beta_{10}} under the null. The
values of \code{L} set to \code{F} are estimated.
}
\examples{
yt = D.bnr$yt;
ys = D.bnr$ys;
Zt = D.bnr$Zt;
Zs = D.bnr$Zs;
# Test for an overall effect, should reject.
Score.bnr(yt,ys,Zt,Zs,L=c(FALSE,TRUE,TRUE,TRUE));
# Test for the effect of \\eqn{\\beta_{3}}, should not reject.
Score.bnr(yt,ys,Zt,Zs,L=c(FALSE,FALSE,FALSE,TRUE));
}
