% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tests.R
\name{Score.bnr}
\alias{Score.bnr}
\title{Score Test}
\usage{
Score.bnr(y.t, y.s, D.t, D.s, L, b0, maxit = 10, eps = 1e-06)
}
\arguments{
\item{y.t}{Target outcome.}

\item{y.s}{Surrogate outcome.}

\item{D.t}{Target design matrix.}

\item{D.s}{Surrogate design matrix.}

\item{L}{Logical vector, with as many entires as columns in the target
design, indicating which columns of the target design are fixed under the
null.}

\item{b0}{Value of the regression coefficient for the selected columns under
the null. Defaults to zero.}

\item{maxit}{Maximum number of parameter updates.}

\item{eps}{Minimum acceptable improvement in log likelihood.}
}
\description{
Score Test
}
\examples{
y.t = D.bvr$y.t;
y.s = D.bvr$y.s;
D.t = D.bvr$D.t;
D.s = D.bvr$D.s;
# Test for an overall effect
Score.bnr(y.t,y.s,D.t,D.s,L=c(TRUE,TRUE,TRUE));
# Test for the effect of \\eqn{\\beta_{3}}
Score.bnr(y.t,y.s,D.t,D.s,L=c(FALSE,FALSE,TRUE));
}
