% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multivariate.R
\name{Score.mnr}
\alias{Score.mnr}
\title{Multivariate Regression Score Test}
\usage{
Score.mnr(y.t, y.s, Z.t, L.s, L, b0, parallel = F, maxit = 10,
  eps = 1e-06)
}
\arguments{
\item{y.t}{Target outcome.}

\item{y.s}{Surrogate outcome.}

\item{Z.t}{Numeric model matrix for the target outcome.}

\item{L.s}{List of numeric model matrices for the surrogate outcomes.}

\item{L}{Logical vector, with as many entires as columns in the target
design, indicating which columns design are fixed under the null.}

\item{b0}{Value of the regression coefficient for the selected columns under
the null. Defaults to zero.}

\item{parallel}{If true, assumes a parallel coefficient design, where each 
secondary outcome has the same regression parameters.}

\item{maxit}{Maximum number of parameter updates.}

\item{eps}{Minimum acceptable improvement in log likelihood.}
}
\description{
Score test of \eqn{H_{0}:\beta_{1}=\beta_{10}} in the multivariate outcome model.
}
\examples{
y.t = D.mnr$y.t;
y.s = D.mnr$y.s;
Z.t = D.mnr$Z.t;
L.s = D.mnr$L.s;
# Test for an overall effect
Score.mnr(y.t,y.s,Z.t,L.s,L=c(FALSE,FALSE,TRUE,TRUE));
# Test for the effect of \\eqn{\\beta_{3}}
Score.mnr(y.t,y.s,Z.t,L.s,L=c(FALSE,FALSE,FALSE,TRUE));
}
