% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiTest.R
\name{Score.mnr}
\alias{Score.mnr}
\title{Multivariate Regression Score Test}
\usage{
Score.mnr(yt, Ys, Zt, Ls, L, b10, parallel = F, maxit = 100, eps = 1e-06,
  report = F)
}
\arguments{
\item{yt}{Target outcome vector.}

\item{Ys}{Surrogate outcome matrix.}

\item{Zt}{Numeric model matrix for the target outcome.}

\item{Ls}{List of numeric model matrices for the surrogate outcomes.}

\item{L}{Logical vector, with as many entires as columns in the target
design, indicating which columns design are fixed under the null.}

\item{b10}{Value of the regression coefficient for the selected columns under
the null. Defaults to zero.}

\item{parallel}{If true, assumes a parallel coefficient design, where each 
secondary outcome has the same regression parameters.}

\item{maxit}{Maximum number of parameter updates.}

\item{eps}{Minimum acceptable improvement in log likelihood.}

\item{report}{Report model fitting progress?}
}
\description{
Score test of \eqn{H_{0}:\beta_{1}=\beta_{10}} in the multivariate outcome
model. The score test is specified using a logical vector \code{L}, with as
many entries as columns in the target model matrix. The values of \code{L} 
set to \code{T} are fixed at \eqn{\beta_{10}} under the null. The values of 
\code{L} set to \code{F} are estimated.
}
\section{Useage notes}{

\itemize{
 \item \code{Ls} is a list of model matrices, one for each surrogate outcome.
 Each matrix should the same number of rows, but the number of columns may
 vary. The overall design matrix is constructed internally. By default, each
 surrogate outcome is allocated its own regression coefficients. If a
 parallel design is specified, then all surrogate outcomes share the same set
 of regression coefficients. In this case, each model matrix in \code{Ls}
 must have the same number of columns as well.
}
}

\examples{
yt = D.mnr$yt;
Ys = D.mnr$Ys;
Zt = D.mnr$Zt;
Ls = D.mnr$Ls;
# Test for an overall effect, should reject.
Score.mnr(yt,Ys,Zt,Ls,L=c(FALSE,TRUE,TRUE,TRUE));
# Test for the effect of \\eqn{\\beta_{3}}, should not reject.
Score.mnr(yt,Ys,Zt,Ls,L=c(FALSE,FALSE,FALSE,TRUE));
}
