% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updateBVR}
\alias{updateBVR}
\title{Bivariate Parameter Update}
\usage{
updateBVR(t, s, X, Xi, B, A, b0, a0, s0)
}
\arguments{
\item{t}{Target outcome}

\item{s}{Surrogate outcome}

\item{X}{Target design}

\item{Xi}{Surrogate design}

\item{B}{Target inc. proj.}

\item{A}{Surrogate inc. proj.}

\item{b0}{Current beta}

\item{a0}{Current alpha}

\item{s0}{Current sigma}
}
\description{
Parameter update for bivariate outcome model.
}
