% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updateMNR}
\alias{updateMNR}
\title{Multivariate Parameter Update}
\usage{
updateMNR(yt, S, Zt, Zs, At, b0, a0, s0)
}
\arguments{
\item{yt}{Target outcome}

\item{S}{Surrogate outcome vector}

\item{Zt}{Target design}

\item{Zs}{Surrogate design}

\item{At}{Target inc. proj.}

\item{b0}{Current beta}

\item{a0}{Current alpha}

\item{s0}{Current sigma}
}
\description{
Parameter update for multivariate outcome model.
}
